Have to leave.. don't have a key to this place

+++ src/BGL/config.h
@@ -8,7 +8,7 @@
 // the input files will be.
 //
 
-typedef float Scalar;
+typedef double Scalar;
 
 #endif // CONFIG_H
